+++ CHANGELOG.md
@@ -6,13 +6,14 @@
 * Remove `cookTree` and `cookNonEmpty`. 
 * Added `nest`.
 * `cook` is now "typed": we pass the type of the bean we want to extract.
+* `RecipeError` -> `CookingError`.
 * Renamed `PrimaryBean` to `FinishedBean`.
 * Renamed `SecondaryBean` to `AggregateBean`.
 * Now the `Constructor`s don't depend directly on `SecondaryBean`/`AggregateBean`.
   There is a `PrimaryBean`/`FinishedBean` that points to the `SecondaryBean`/`AggregateBean`,
   and `Constructor`s depend on that.
 * Rename `collapseToPrimaryBeans` to `collapseBeans`.
-* Rename `removeSecodnaryBeans` to `removeAggregates`.
+* Rename `removeSecondaryBeans` to `removeAggregates`.
 
 ## 0.6.1.0